Refactor..
