+++ python/nav/django/context_processors.py
@@ -23,7 +23,7 @@ from operator import attrgetter
 from django.conf import settings
 
 from nav.config import find_configfile
-from nav.django.auth import get_sudoer
+from nav.web.auth import get_sudoer
 from nav.django.utils import get_account, is_admin
 from nav.web.message import Messages
 from nav.web.webfront.utils import tool_list, quick_read, split_tools
