+++ python/nav/django/settings.py
@@ -126,8 +126,8 @@ TEMPLATES = [
 MIDDLEWARE = (
     'django.middleware.common.CommonMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
-    'nav.django.auth.AuthenticationMiddleware',
-    'nav.django.auth.AuthorizationMiddleware',
+    'nav.web.auth.AuthenticationMiddleware',
+    'nav.web.auth.AuthorizationMiddleware',
     'nav.django.legacy.LegacyCleanupMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
 )
